+++ apps/docs/stories/Button.stories.tsx
@@ -1,41 +1,35 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { StoryObj } from '@storybook/react';
 import { Button } from '@najwer23snacks/button';
 
-const meta: Meta<typeof Button> = {
-  title: 'Buttons/Button',
+const meta = {
+  title: 'Buttons/Button Classic',
   component: Button,
-  argTypes: {
-    type: {
-      control: { type: 'radio' },
-      options: ['button', 'submit', 'reset'],
-    },
+  parameters: {
+    layout: 'centered',
   },
   tags: ['autodocs'],
 };
 
 export default meta;
 
-type Story = StoryObj<typeof Button>;
+type Story = StoryObj<typeof meta>;
 
-/*
- *👇 Render functions are a framework specific feature to allow you control on how the component renders.
- * See https://storybook.js.org/docs/react/api/csf
- * to learn how to use render functions.
- */
-export const Primary: Story = {
-  render: (props) => (
-    <Button
-      {...props}
-      onClick={(): void => {
-        // eslint-disable-next-line no-alert -- alert for demo
-        alert('Hello from Turborepo!');
-      }}>
-      Hello
-    </Button>
-  ),
-  name: 'Button',
+export const Default: Story = {
   args: {
-    children: 'Hello',
-    type: 'button',
+    children: 'Click!',
+    disabled: false,
+    onClick: () => {
+      console.log('Hello Click!');
+    },
+  },
+};
+
+export const Disabled: Story = {
+  args: {
+    children: 'Click!',
+    disabled: true,
+    onClick: () => {
+      console.log('Hello Click!');
+    },
   },
 };