+++ packages/button/src/button.tsx
@@ -1,13 +1,16 @@
 import styles from './Button.module.css';
 
-export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  children: React.ReactNode;
-}
-
-export const Button = ({ children, ...other }: ButtonProps): JSX.Element => {
+export const Button: React.FC<
+  React.ButtonHTMLAttributes<HTMLButtonElement> & {
+    children: React.ReactNode;
+    kind?: 'classic';
+  }
+> = ({ children, kind = 'classic', ...props }): JSX.Element => {
   return (
-    <button type="button" {...other} className={styles.test}>
+    <button type="button" {...props} className={styles[kind]}>
       {children}
     </button>
   );
 };
+
+Button.displayName = 'Button';