+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # docs
 
+## 1.2.0
+
+### Minor Changes
+
+- Classic Button with CSS tokens
+
 ## 1.1.0
 
 ### Minor Changes